fix python3 warnings

+++ ldk/drivers/oscillo.py
@@ -137,11 +137,11 @@ class Oscillo(object):
         self.spectrum = fft_adc[:, 0:self.wfm_size / 2]
 
     def get_avg_spectrum(self, n_avg=1):
-        self.avg_spectrum = np.zeros((2, self.wfm_size / 2))
+        self.avg_spectrum = np.zeros((2, int(self.wfm_size / 2)))
         for i in range(n_avg):
             self.get_adc()
             fft_adc = np.abs(np.fft.fft(self.adc, axis=1))
-            self.avg_spectrum += fft_adc[:, 0:self.wfm_size / 2]
+            self.avg_spectrum += fft_adc[:, 0:int(self.wfm_size / 2)]
         self.avg_spectrum /= n_avg
 
     # -------------------------------
